Pass our new CL options on to importers
+ Options pass through district-level settings and then to importers
+ Break district settings into separate files and test 'em
